Implementation of bouncing off of left paddle
This was somewhat of a difficult task, as the states have to be kept in the correct scope. For example the useEffect has to keep track of the correct states, which has to be defined carefully. Another pitfall was the time at which changes were realized. This was not always when I wanted and needed them to be. This solution now makes use of a constant speed variable, that doesnt change at runtime but can still define the speed at which the game is played.