Added randomness and difficultylevels adjusted
The difficultylevel now determines the level of randomness in the angle of the bounce of the ball, as well as the speed of the bounce. This can cause quite difficult game play scenarios. The length of the paddle as well as the speed of it is reduced to increase difficulty.