Adjustments and tickets
Included a Profile page fetch of the user Achievements.

Included statistics detail page for the profile page.

fixed a bug, where an achievement would be awarded twice for the same game.

made the return button on the detail pages static, so that you dont have to scroll all the way back to the top to exit that view.